extract use-with-cancel
