+++ application/ui/src/features/annotator/tools/segment-anything-tool/segment-anything-tool.component.tsx
@@ -19,6 +19,7 @@ import { SAMLoading } from './sam-loading.component';
 import { InteractiveAnnotationPoint } from './segment-anything.interface';
 import { useSegmentAnythingModel } from './use-segment-anything.hook';
 import { useSingleStackFn } from './use-single-stack-fn.hook';
+import { useWithCancel } from './use-with-cancel';
 
 import classes from './segment-anything.module.scss';
 
@@ -49,44 +50,6 @@ const PreviewAnnotations = ({ previewAnnotations, image }: PreviewAnnotationsPro
             ))}
         </MaskAnnotations>
     );
-};
-
-const useWithCancel = (fn: (points: InteractiveAnnotationPoint[]) => Promise<Shape[]>) => {
-    const abortController = useRef<AbortController | null>(null);
-
-    const cancellableCallback = useCallback(
-        async (...args: Parameters<typeof fn>) => {
-            // Cancel any ongoing request
-            abortController.current?.abort();
-
-            abortController.current = new AbortController();
-
-            const result = await fn(...args);
-
-            if (abortController.current.signal.aborted) {
-                throw new DOMException('Request aborted', 'AbortError');
-            }
-
-            return result;
-        },
-        [fn]
-    );
-
-    const cancel = useCallback(() => {
-        abortController.current?.abort();
-        abortController.current = null;
-    }, [abortController]);
-
-    useEffect(() => {
-        return () => {
-            cancel();
-        };
-    }, [cancel]);
-
-    return {
-        call: cancellableCallback,
-        cancel,
-    };
 };
 
 export const SegmentAnythingTool = () => {
